pruning logic and comprehensive tests in
